% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculus.R
\name{calculate_rank}
\alias{calculate_rank}
\title{Calculate Rankings}
\usage{
calculate_rank(
  home_team = NULL,
  away_team = NULL,
  score = c(0, 0),
  world_cup = FALSE
)
}
\arguments{
\item{home_team}{Character vector of home team name. For a list of valid teams,
see \code{teams()}.}

\item{away_team}{Character vector of away team name.}

\item{score}{Two-element numeric vector of scores.}

\item{world_cup}{Whether the match took place in the Rugby World Cup or not.}
}
\description{
Calculate rankings based on a match between two teams, the score,
home advantage and whether or not the match took place in a World Cup.
}
