% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/espn_team.R
\name{get_team_records}
\alias{get_team_records}
\title{Get Match Data for Rugby Teams}
\usage{
get_team_records(country = NULL, all = TRUE)
}
\arguments{
\item{country}{Name of country.}

\item{all}{\code{TRUE} by default, returns info on all teams.}
}
\description{
Download data from ESPN's statsguru service. Data includes information
on various teams, win ratios, points scored etc.
}
